+++ gateway/src/apicast/module.lua
@@ -32,7 +32,7 @@ local prequire = function(file)
   return ok, ret
 end
 
-local name = env.get('APICAST_MODULE') or 'apicast'
+local name = env.get('APICAST_MODULE') or 'apicast.policy.apicast'
 
 local ok, mod = prequire(name)
