+++ t/apicast-policy-chains.t
@@ -26,7 +26,7 @@ defined at the service level.
           backend_authentication_type = 'service_token',
           backend_authentication_value = 'token-value',
           proxy = {
-            policy_chain = { { name = 'apicast.policy.phase_logger' }, { name = 'apicast' } },
+            policy_chain = { { name = 'apicast.policy.phase_logger' }, { name = 'apicast.policy.apicast' } },
             api_backend = "http://127.0.0.1:$TEST_NGINX_SERVER_PORT/api-backend/",
             proxy_rules = {
               {
